bookorbooks/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-filter==2.4.0
django-nested-inline==0.4.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
Markdown==3.3.4
Pillow==8.3.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
yapf==0.31.0
